Updated gitignore.

+++ utilities/schema.go
@@ -1,12 +1,12 @@
 package utilities
 
 type UserType struct {
-	ID   int    `json:"id"`
+	ID   int64  `json:"id"`
 	Name string `json:"name"`
 }
 
 type User struct {
-	ID       int    `json:"id"`
+	ID       int64  `json:"id"`
 	Name     string `json:"name"`
 	UserType int    `json:"user_type"`
 	Email    string `json:"email"`
@@ -14,7 +14,7 @@ type User struct {
 }
 
 type RoomType struct {
-	ID    int     `json:"id"`
+	ID    int64   `json:"id"`
 	Name  string  `json:"name"`
 	Beds  int     `json:"beds"`
 	Beach bool    `json:"beach"`
@@ -22,7 +22,7 @@ type RoomType struct {
 }
 
 type Room struct {
-	ID       int    `json:"id"`
+	ID       int64  `json:"id"`
 	Number   int    `json:"number"`
 	RoomType int    `json:"room_type"`
 	Floor    int    `json:"floor"`
@@ -30,7 +30,7 @@ type Room struct {
 }
 
 type Booking struct {
-	ID     int    `json:"id"`
+	ID     int64  `json:"id"`
 	RoomID int    `json:"room_id"`
 	UserID int    `json:"user_id"`
 	Start  string `json:"start"`
